requirements.txt
Django==4.0.3
djangorestframework==3.11.0
requests==2.25.1